+++ autogalaxy/aggregator/__init__.py
@@ -3,4 +3,4 @@ from autogalaxy.aggregator.imaging import ImagingAgg
 from autogalaxy.aggregator.fit_imaging import FitImagingAgg
 from autogalaxy.aggregator.interferometer import InterferometerAgg
 from autogalaxy.aggregator.fit_interferometer import FitInterferometerAgg
-from autogalaxy.ellipse.aggregator.ellipses import EllipsesAgg
+from autogalaxy.aggregator.ellipses import EllipsesAgg
