+++ autogalaxy/aggregator/ellipses.py
