+++ test_autogalaxy/aggregator/ellipse/conftest.py
