+++ test_autogalaxy/aggregator/ellipse/test_aggregator_ellipses.py
